changing database and adding option to update task still dev

+++ app/src/main/java/pl/mazy/todoapp/logic/dataClass/Task.kt
@@ -1,7 +1,7 @@
 package pl.mazy.todoapp.logic.dataClass
 
-data class Task (
-    val ID:Int,
+data class Task(
+    val ID: Long,
     val name:String,
     val checked:Boolean,
     val category:String
